move toolbar to dedicated file
